% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidatos.R
\name{candidatos}
\alias{candidatos}
\title{Busca los candidatos dentro de nuestro archivo de anotaciones de F. vesca}
\usage{
candidatos(
  path = getwd(),
  distance_SNPs = 350,
  min.P.value = 1e-07,
  rm.duplicados = TRUE
)
}
\arguments{
\item{path}{Ruta de la carpeta con los documentos}

\item{distance_SNPs}{Numero de pb a desplazarse hacia ambos extremos comenzando en la localizacion del SNP}

\item{min.P.value}{El minimo valor de P.value usado como limite para seleccionar los SNPs. Por defecto, 1e-07}

\item{rm.duplicados}{Parametro que determina si se eliminan los genes duplicados o no}
}
\value{
Tras correr este codigo se genera un .xlsx en la carpeta Resultados, en el que se muestran los genes candidatos con los criterios seleccionados
}
\description{
Selecciona aquellos candidatos en funcion del numero de SNPs relevantes y la distancia hacia ambos extremos
}
