% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidatos_old_version.R
\name{candidatos_old_version}
\alias{candidatos_old_version}
\title{Busca los candidatos dentro de nuestro archivo de anotaciones de F. vesca}
\usage{
candidatos_old_version(input, number_SNPs, distance_SNPs = 350000)
}
\arguments{
\item{input}{Nuestra tabla obtenida para un trait después de correr GAPIT}

\item{number_SNPs}{Número de SNPs con menor Pvalue a tener en cuenta. Por lo general, se selecciona mirando el Manhattan plot}

\item{distance_SNPs}{Número de pb a desplazarse hacia ambos extremos comenzando en la localización del SNP}
}
\value{
Tras correr este codigo se genera un .xlsx en la carpeta Resultados, en el que se muestran los genes candidatos con los criterios seleccionados
}
\description{
Selecciona aquellos candidatos en función del número de SNPs relevantes y la distancia hacia ambos extremos
}
