% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clasificar_helsinki.R
\name{clasificar_helsinki}
\alias{clasificar_helsinki}
\title{Clasifica las lecturas en funcion del Patron de Genotipos (Coleccion Helsinki)}
\usage{
clasificar_helsinki(samplelist)
}
\value{
Tras correr este codigo se genera un .txt que es directamente guardado en la carpeta en la que nos encontremos
}
\description{
Añade el numero de correspondencia de cada genotipo en la 2 columna. Genera un .txt que puede ser empleado como samplelist en TagFinder
}
