% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heatmaps}
\alias{heatmaps}
\title{Genera un heatmap normalizando los datos por la mediana}
\usage{
heatmaps(
  datos,
  barra_secundaria = TRUE,
  barra_secundaria_nombre = "Location",
  reorganizar_Col = TRUE,
  modelo_orden = "dist"
)
}
\arguments{
\item{datos}{La lista con las lecturas correspondiente a cada genotipo}

\item{barra_secundaria}{Permite activar o desactivar la barra secundaria a mostrar en el heatmap}

\item{barra_secundaria_nombre}{Lo que significa esta barra secundaria (Genotipo, localizacion...)}

\item{modelo_orden}{Función a usar para clusterizar. Por defecto es "dist", pero se puede seleccionar entre "pearson", "spearman" y "kendall"}

\item{reorganizar_col}{Permita activar o desactivar el clustering de columnas}
}
\value{
Tras correr este código se muestra el Heatmap en la ventana "Viewer".
}
\description{
Función que genera un heatmap de nuestros datos. Para que funcione, hay que importar los datos de manera que la primera fila sea nuestra variable independiente (genotipos), y la primera columna sea el nombre de nuestros metabolitos
}
