\name{metabolomicsChanges}
\alias{metabolomicsChanges}

\title{Changes made to the metabolomics package}
\description{Write first \code{n} lines of the recent changes made to the metabolomics package}

\usage{
    metabolomicsChanges(n = 5)
}

\arguments{
    \item{n}{The number of lines to write of changelog.}
}

\value{
    Text with first \code{n} lines of the recent changes made to the metabolomics package.
}

\author{Alysha M De Livera, Jairus B Bowne}

\examples{
    metabolomicsChanges(8)
}
