% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primarios.R
\name{primarios}
\alias{primarios}
\title{Selecciona las mejores lecturas para un set analizado por TagFinder}
\usage{
primarios(set, tiempos = Tiempos_Especificos_Fruto)
}
\arguments{
\item{set}{.txt generado por tag finder}

\item{tiempos}{Plantilla de Tiempos Especificos. Por defecto hay 3 (Hoja, Fruto y Split)(Todo de fresa)}
}
\value{
Tras correr este código se presentan los metabolitos que faltan por encontrar. Incluir los tiempos especificos en TagFinder para un segundo filtro
}
\description{
Función que genera una lista final de los metabolitos seleccionados (Plantilla Tiempos Especificos)
}
