% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primarios_faltan.R
\name{primarios_faltan}
\alias{primarios_faltan}
\title{Selecciona las mejores lecturas para un set analizado por TagFinder}
\usage{
primarios_faltan(set2, set_primarios, tiempos = Tiempos_Especificos_Fruto)
}
\arguments{
\item{set2}{.txt generado por tag finder aplicando los filtros especificos de la funcion "primarios}

\item{set_primarios}{lista generada por la funcion "primarios". Se usa como input en esta funcion}

\item{tiempos}{Plantilla de Tiempos Especificos. Por defecto hay 3 (Hoja, Fruto y Split)(Todo de fresa)}
}
\value{
Tras correr este código se presentan los metabolitos que faltan por encontrar.
}
\description{
Función que genera una lista final de los metabolitos seleccionados (Plantilla Tiempos Especificos)
}
