\name{treated}
\alias{treated}
\title{A metabolomics study with paired observations.}
\description{A data frame with data collected from several subjects before 
    and after a specific treatment.}

\usage{data(treated)}

\keyword{datasets}

\author{ Alysha M De Livera, Jairus B Bowne}

\examples{
    data(treated)
    treated
}
